do builder more portable separate posix tools from builder build do cdsay internal builder command

+++ bld/builder/c/sysdoscd.c
@@ -37,7 +37,7 @@
 int SysDosChdir( char *dir )
 {
     char        *end;
-    unsigned    len;
+    size_t      len;
     unsigned    total;
 
     if( dir[0] == '\0' )
